Separate client and server test configurations
In order to get this to work properly, I had to run the client and
server tests separately.  This shouldn't be required because since
version 20, Jest supports running multiple "projects" with different
configurations in a single runner.  However, I was unable to get it to
work as expected.  Some of the problems I encountered while
experimenting with different setups include:

* each test run twice
* global configuration could not be specified from a file
* when specifying global configuration from the command line, I had to
  explicitly override rootDir